matching new API, decoupling model and  train module
